feat: merge autocomplete disabling test that are failing
